requirements.txt
pymodsecurity==0.1.0
typer
